requirements-dev.txt
pytest==2.9.2
pytest-django==2.9.1
pytest-cov
pytest-xdist
flake8
responses
beautifulsoup4
